pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool]
[tool.black]
line-length = 120
target-version = ["py37", "py38"]
include = '\.pyi?$'
exclude = '''
(?x)(
    \.git
  | build
  | _build
  | dist
  | docs

  | pyroclast/schedule\.py
  | tests/test_schedule\.py
)
'''

[tool.poetry]
name = "pyroclast"
version = "0.1.0"
description = "Data Platform Integration Library"
authors = ["Eric Lee <elee@kududyn.com>"]
readme = "README.md"
repository = "https://github.com/kudu/pyroclast"
homepage = "https://github.com/kudu/pyroclast"
classifiers = [
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.7",
    "Private :: Do Not Upload",
]

[tool.poetry.dependencies]
namesgenerator = "==0.3"
python = "==3.*,>=3.7.0"
python-nomad = "^1.1"
tenacity = "==5.*"
asyncio-nats-client = {version = "^0.9.2",optional = true}
asyncio-nats-streaming = {version = "^0.4.0",optional = true}
pydgraph = {version = "^2.0.2",optional = true}
redis = {version = "^3.3",optional = true}
boto3 = {version = "^1.10",optional = true}
asonic = {version = "^1.0.3",optional = true}

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-black = "*"
pytest-cov = "*"
pytest-flake8 = "*"

[tool.poetry.extras]
dgraph = ["pydgraph"]
nats = ["asyncio-nats-client", "asyncio-nats-streaming"]
redis = ["redis"]
s3 = ["boto3"]
sonic = ["asonic"]

[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}
